requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
PyJWT==2.8.0
Werkzeug==3.0.1
watchdog==3.0.0 